+++ CHANGELOG.md
@@ -7,3 +7,5 @@
 - Add Factory\Controller\Plugin\UserSettingPluginFactory test
 - Add Factory\Service\UserSettingsServiceFactory test
 - Add Factory\View\Helper\UserSettingHelperFactory test
+- Changed passed Service Name in Mapper\DoctrineORM\UserSettingMapperFactory as 'Doctrine\ORM\EntityManager' as there is no service named 'zfcuser_doctrine_em'
+- Add Factory\Mapper\DoctrineORM\UserSettingMapperFactory test